BE/flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
redis==5.0.1
python-dotenv==1.0.0
email-validator==2.1.0
Pillow==10.2.0
pytest==7.4.4
pytest-flask==1.3.0